Adding changes to csexercise1

+++ csexercise1.py
@@ -1,12 +1,12 @@
-hours = int(input('Enter Hours:'))
-rate = int(input('Enter Rate:'))
+hours = float(input('Enter Hours:'))
+rate = float(input('Enter Rate:'))
 
 if hours > 40:
-    overtime = float(hours) - 40
+    overtime = hours - 40
 else :
     overtime = 0
 
-overtimepay = 0.5 * float(rate) * overtime
-pay = float(hours) * float(rate) + overtimepay
+overtimepay = 0.5 * rate * overtime
+pay = hours * rate + overtimepay
 
 print('Pay:' + str(pay))
